Push debug symbols to vice
Small hacky python script to pull the symbol table out of the vasm
listing file and turn it into add_label commands that vice understands.
And I've been building with a small wrapper that calls the python script
after the assembler is done and uses netcat to tell vice to reload the
symbols from the file.